Integration from branch Visualization_Library_SDK-2011-NG.
- App_TessellationShader made (almost) compatible with OpenGL 4 Core.
- VertexAttribInfo construction parameters order changed and added defaults.
- Added method toGenericVertexAttribs() to convert a Geometry vertex array to generic ones.

GLSL
- Log warning on GLSLProgram linkage in debug mode.
- methods *AutomaticAttribLocation() renamed *AutoAttribLocation()
- added getUniformLocation() taking const char* as argument.
- added activeUniformLocations() method to access the list of currently active uniforms and relative binding locations.
- added methods vl_ModelViewMatrix(), vl_ProjectionMatrix(), vl_ModelViewProjectionMatrix() and vl_NormalMatrix() to access VL built-in uniforms locations and test for their existence after GLSLProgram linking.

OpenGL 4
- Experimental OpenGL 4 Core support!
- Unified ProjViewTransfCallback class can automatically update OpenGL 1.x matrices as well as VL built-in matrices for GL 4 Core profile: vl_ModelViewMatrix, vl_ProjectionMatrix, vl_ModelViewProjectionMatrix, vl_NormalMatrix.
- OpenGL 4 Core mipmaps generation support.

+++ src/vlGraphics/Terrain.cpp
@@ -183,7 +183,7 @@ void Terrain::init()
 
     glsl = new GLSLProgram;
     ref<Uniform> Height = new vl::Uniform("Height");
-    Height->setUniformF(height());
+    Height->setUniformF((float)height());
     glsl->setUniform( Height.get() );
     glsl->attachShader( new GLSLFragmentShader( String::loadText(fragmentShader()) ) );
     if (use_uniform_transform)
